Django_Ideal/requirements.txt
pandas
transformers==4.5.1
spacy==3.0.6
tensorflow==2.2.0
django ==3.2
django-rest-framework
django_tables2
celery==4.4.2